+++ src/big_luca/config.rs
@@ -7,6 +7,7 @@
 pub struct Config {
     pub database_url: String,
     pub redis_url: String,
+    pub rsshub_url: String,
     pub teloxide_token: String,
 }