+++ src/main.rs
@@ -12,6 +12,7 @@ mod big_luca;
 mod feed;
 mod redis;
 mod repository;
+mod rsshub;
 mod utils;
 mod youtube;